requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
pandas>=2.0.0
plotly>=5.18.0
gunicorn>=21.2.0
python-Levenshtein>=0.21.0
Faker>=22.0.0
django-ledger>=0.6.3
django-ninja>=1.1.0
